More example narrowing warning fixes

+++ examples/MIDIUSB_write/MIDIUSB_write.ino
@@ -34,7 +34,7 @@ void setup() {
 // Fourth parameter is the control value (0-127).
 
 void controlChange(byte channel, byte control, byte value) {
-  midiEventPacket_t event = {0x0B, 0xB0 | channel, control, value};
+  midiEventPacket_t event = {0x0B, (uint8_t)(0xB0 | channel), control, value};
   MidiUSB.sendMIDI(event);
 }
